Bumped up to TestNG 6.2
For service loader lookup for test selector to work, customers need to
use TestNG 6.2 or later, which contains this crucial bug fix: https://github.com/testng-team/testng/commit/f38bbfab2b620883b38039baac48585efd057bf3

So I decided that it'd be prudent for us to test this library with that
version.

The new version of TestNG uncovered a bug in our code, where we were
modifying the list given to us. TestNG apparently doesn't expect this in
some cases, so we needed to make a defensive copy.

With this change, I also needed to update tests, in part because TestNG
is now passing the same methods multiple times. That means previous
naive counting approach doesn't work anymore.

I also refactored the test code somewhat to make it more DRY.